QOL: Save warning, rename tabs
Add a warning to ask if the user wants to save the current tab before
closing and losing progress. If the user confirms that they will save
the file, but clicks 'Cancel' on the Windows Prompt when attepting to
save, the program will reprompt them to avoid unwanted progress loss.
Notepad tab titles will update based on the chosen file name upon
saving.